fix: groups object circular dependency
move updatePollingRelease() to a different file
